chore(ava): delete useless code and move prompt to same folder
