+++ src/prompt/extract.ts
